🗃Refactor: setHtml clearHtml 상위 컴포넌트로 이동
