+++ homework/dhrod0325/03/src/components/Death/DeathTotal.ts
@@ -1,14 +1,10 @@
 import { Country } from 'covid';
 import { BaseComponent } from '@/lib/Component';
-import { SummaryWrapper } from '@/@model/SummaryWrapper';
+import { SummaryWrapper } from '@/model/SummaryWrapper';
 
 export class DeathTotal extends BaseComponent {
   public loadData(summaryWrapper: SummaryWrapper) {
     this.setHtml(`${summaryWrapper.getTotalDeaths()}`);
-  }
-
-  public setHtml(count: string) {
-    this.getContainer().innerText = count;
   }
 
   public setHtmlByFirstCountry(data: Country[] | undefined) {
