+++ homework/dhrod0325/03/src/components/LastUpdate/LastUpdate.ts
@@ -1,14 +1,10 @@
 import { getDateString } from '@/lib/utils';
 import { BaseComponent } from '@/lib/Component';
-import { SummaryWrapper } from '@/@model/SummaryWrapper';
+import { SummaryWrapper } from '@/model/SummaryWrapper';
 
 export class LastUpdate extends BaseComponent {
   public setup(data: SummaryWrapper): void {
     this.update(data);
-  }
-
-  public setHtml(html: string): void {
-    this.getContainer().innerText = html;
   }
 
   private update(data: SummaryWrapper): void {
