+++ homework/dhrod0325/03/src/components/Recovered/RecoveredTotalList.ts
@@ -3,7 +3,7 @@ import { RecoveredTotal } from './RecoveredTotal';
 import { api } from '@/lib/Api';
 import { DefaultSpinner } from '@/components/Helper/DefaultSpinner';
 import { AsyncComponent } from '@/lib/Component';
-import { SummaryWrapper } from '@/@model/SummaryWrapper';
+import { SummaryWrapper } from '@/model/SummaryWrapper';
 
 export class RecoveredTotalList extends AsyncComponent {
   private readonly SPINNER_ID = 'recovered-spinner';
@@ -23,7 +23,7 @@ export class RecoveredTotalList extends AsyncComponent {
   }
 
   public prepareAsync(): void {
-    this.$list.clear();
+    this.$list.clearHtml();
   }
 
   public async loadDataAsync(selectedId: string) {
