+++ homework/dhrod0325/03/src/model/SummaryWrapper.ts
